Splitted single django_utils.py to tech sliced modules
