+++ cvat/apps/engine/view_utils.py
@@ -6,14 +6,12 @@
 
 import textwrap
 from datetime import datetime
-from typing import Any
 
-from django.db.models import Manager, QuerySet
+from django.db.models import QuerySet
 from django.http import HttpResponseGone
 from django.http.response import HttpResponse
 from drf_spectacular.utils import extend_schema
 from rest_framework.decorators import action
-from rest_framework.exceptions import NotFound
 from rest_framework.response import Response
 from rest_framework.serializers import Serializer
 from rest_framework.viewsets import GenericViewSet
@@ -21,7 +19,6 @@ from rest_framework.viewsets import GenericViewSet
 from cvat.apps.engine.parsers import TusUploadParser
 from cvat.apps.engine.tus import TusFile
 from cvat.apps.engine.types import ExtendedRequest
-from cvat.utils.django_database import _ModelT
 
 
 def make_paginated_response(
@@ -125,24 +122,3 @@ def deprecate_response(response: Response, *, deprecation_date: datetime) -> Non
     # https://www.rfc-editor.org/rfc/rfc9745
     deprecation_timestamp = int(deprecation_date.timestamp())
     response.headers["Deprecation"] = f"@{deprecation_timestamp}"
-
-
-def get_or_404(
-    queryset: type[_ModelT] | QuerySet[_ModelT] | Manager[_ModelT],
-    pk: Any,
-) -> _ModelT:
-    """
-    A simpler version of django.shortcuts.get_object_or_404()
-    Produces a better error message.
-    """
-
-    if hasattr(queryset, "_default_manager"):
-        queryset = queryset._default_manager.all()
-
-    model_type = queryset.model
-
-    try:
-        return queryset.get(pk=pk)
-    except model_type.DoesNotExist as ex:
-        readable_model_name = queryset.model._meta.verbose_name.capitalize()
-        raise NotFound(f"{readable_model_name} {pk} does not exist") from ex
