+++ cvat/apps/events/handlers.py
@@ -47,7 +47,7 @@ from cvat.apps.organizations.serializers import (
 )
 from cvat.apps.webhooks.models import Webhook
 from cvat.apps.webhooks.serializers import WebhookReadSerializer
-from cvat.utils.django_database import find_psycopg_cause
+from cvat.utils.django_database.utils import find_psycopg_cause
 
 from .cache import get_cache
 from .const import WORKING_TIME_RESOLUTION, WORKING_TIME_SCOPE
